在办接口规范化处理
(cherry picked from commit e5230d29b16ffccf770f427bddaa7a4f7259c020)

+++ y9-module-processadmin/risenet-y9boot-support-processadmin/src/main/java/net/risesoft/api/DoingApiImpl.java
@@ -132,7 +132,7 @@ public class DoingApiImpl implements DoingApi {
 
         totalCount = (int)historyService.createNativeHistoricProcessInstanceQuery().sql(countSql)
             .parameter("processDefinitionKey", processDefinitionKey + "%").parameter("USER_ID_", userId).count();
-        return Y9Page.success(page, (int)(totalCount + rows - 1) / rows, totalCount, resList);
+        return Y9Page.success(page, (totalCount + rows - 1) / rows, totalCount, resList);
     }
 
     /**
